course-service/requirements.txt
# ============================================================
# course-service/requirements.txt
# Production dependencies for the Course Service
# Python 3.11+  |  FastAPI + SQLAlchemy 2 + asyncpg
# ============================================================

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database — async PostgreSQL
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
greenlet==3.0.3

# JWT validation (shared secret with Login Service)
PyJWT==2.8.0
cryptography==42.0.8

# Config management
pydantic==2.7.1
pydantic-settings==2.3.0

# HTTP client (for inter-service calls if needed)
httpx==0.27.0

# Logging
python-json-logger==2.0.7
